poplar: move pointer/dma pointer macros to "poplar.h"
These macros are defined in "drivers/net/higmacv300/higmac.c", to be
explicit about converting between (possibly 64-bit) pointers and
(32-bit) integer values used as addresses:
    ptr_to_dma32(p)
    dma32_to_ptr(da)

A USB driver that's about to be added also requires this capability,
so move the definition into the common "poplar.h" instead.

Rename the macros to this:
    ptr_to_a32(p)
    a32_to_ptr(a32)
because they generally speaking are operating on 32-bit integers
used as addresses (e.g., of registers) which are not necessarily
DMA addresses.

Signed-off-by: Alex Elder <elder@linaro.org>