poplar: poplar_udc: use a32_to_ptr() in two spots
Two spots in hiusb_write_empty_tx_fifo() read an endpoint's TX status
register, and cast an offset to a pointer, like this:
    d32 = readl((uint32_t volatile *)(REG_DTXFSTS0 + epnum * 0x20));

The cast intends to do two things:  convert the integral register
offset into a pointer; and ensure that pointer is known to refer to
volatile data.

64-bit ARM has 64-bit pointers, so the cast above produces:
    warning: cast to pointer from integer of different size

Address this by making the size conversion explicit by using the
a32_to_ptr() macro.  The expansion of readl() already treats its
argument as pointing to volatile data, so the result should be
just as originally intended.

Signed-off-by: Alex Elder <elder@linaro.org>